+++ .github/workflows/apple-devices.yml
@@ -1,30 +1,32 @@
-name: Uniffi Kotlin Package Tests
+# based on https://gist.github.com/learnopengles/60c945b92fbe60a0fecaa0143c35c4a for testing in simulator
+name: Apple Simulator devices
 
 on: pull_request
 
 env:
-  CARGO_TERM_COLOR: always
   TOOLCHAIN: nightly
+  CARGO_TERM_COLOR: always
+  CARGO_DINGHY_REV: "a8ec5b0"
 
 jobs:
   integration_tests:
-    runs-on: macos-14
-
+    # rust cross-compilation
+    runs-on: macos-13
+    strategy:
+      matrix:
+        rust-target: ["x86_64-apple-ios" ]
+        apple-sim: [ "iOS" ]
+        include:
+          - rust-target: "x86_64-apple-tvos"
+            apple-sim: "tvOS"
+          - rust-target: "x86_64-apple-watchos-sim"
+            apple-sim: "watchOS"
     steps:
       - uses: actions/checkout@v4
-      - name: Install Rust Nightly
-        uses: dtolnay/rust-toolchain@master
+      - uses: dtolnay/rust-toolchain@master
         with:
           toolchain: ${{ env.TOOLCHAIN }}
-          targets:
-            armv7-linux-androideabi
-            aarch64-linux-android
-            i686-linux-android
-            x86_64-linux-android
-            x86_64-unknown-linux-gnu
-            aarch64-apple-darwin
-            x86_64-apple-darwin
-
+          components: rust-src
       - name: Cache
         uses: actions/cache@v4
         with:
@@ -33,7 +35,12 @@ jobs:
             ~/.cargo/git
             ~/.cargo/registry
             target
-          key: ${{ github.workflow }}-${{ github.job }}-toolchain-${{ env.TOOLCHAIN }}-cargo-${{ hashFiles('**/Cargo.lock') }}
+          key: ${{ github.workflow }}-${{ github.job }}-toolchain-${{ env.TOOLCHAIN }}-cargo-${{ env.CARGO_DINGHY_REV }}-${{ hashFiles('**/Cargo.lock') }}
+      - name: Install cargo-dinghy
+        run: |
+          if ! command -v cargo-dinghy &> /dev/null; then
+            cargo install --git https://github.com/liveview-native/dinghy.git --force --rev ${CARGO_DINGHY_REV} cargo-dinghy
+          fi
 
       - name: Cache deps
         id: cache-deps
@@ -66,22 +73,16 @@ jobs:
           mix compile
           mix phx.server &
 
-      - name: Set up JDK 17
-        uses: actions/setup-java@v4
+      - name: Start Simulator
+        uses: futureware-tech/simulator-action@v3
         with:
-          java-version: '17'
-          distribution: 'temurin'
+          os: ${{ matrix.apple-sim }}
 
-      - name: Install gradle
-        uses: gradle/gradle-build-action@v2.11.0
-
-      # https://github.com/actions/runner-images/blob/5a9870800dc54127d22ee36c303b47314d231c69/images/macos/macos-14-arm64-Readme.md
-      # There's no python in the m1 macos runners.
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-
-      - name: Run build with Gradle wrapper
-        working-directory: ./crates/core/liveview-native-core-jetpack/
-        run:
-          RANLIB=${ANDROID_NDK_HOME}/toolchains/llvm/prebuilt/darwin-x86_64/bin/llvm-ranlib ./gradlew build --scan
+      - name: Run Rust tests on simulator for ${{ matrix.rust-target }}
+        env:
+          DINGHY_LOG: debug
+        timeout-minutes: 20
+        run: |
+          cargo dinghy all-platforms
+          cargo dinghy all-devices
+          cargo -Zbuild-std test --target ${{ matrix.rust-target }}
