+++ .github/workflows/uniffi-swift-test.yml
@@ -9,7 +9,7 @@ env:
 
 jobs:
   integration_tests:
-    runs-on: macos-13 # This is the beta runner
+    runs-on: macos-14
 
     steps:
       - uses: actions/checkout@v4
@@ -25,12 +25,43 @@ jobs:
             aarch64-apple-darwin
             x86_64-apple-darwin
 
+      - name: Cache deps
+        id: cache-deps
+        uses: actions/cache@v4
+        env:
+          cache-name: cache-elixir-deps
+        with:
+          path: tests/support/test_server/deps
+          key: ${{ runner.os }}-mix-${{ env.cache-name }}-2-${{ hashFiles('**/mix.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-mix-${{ env.cache-name }}-
+      - name: Cache compiled build
+        id: cache-build
+        uses: actions/cache@v4
+        env:
+          cache-name: cache-compiled-build
+        with:
+          path: tests/support/test_server/_build
+          key: ${{ runner.os }}-mix-${{ env.cache-name }}-2-${{ hashFiles('**/mix.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-mix-${{ env.cache-name }}-
+            ${{ runner.os }}-mix-
+      - name: Set up Elixir
+        run: brew install elixir
+
+      - name: Install test server dependencies, compile and run in background
+        working-directory: ./tests/support/test_server
+        run: |
+          mix deps.get
+          mix compile
+          mix phx.server &
+
       - uses: taiki-e/install-action@v2
         with:
           tool: cargo-make
 
       - name: Cache
-        uses: actions/cache@v3
+        uses: actions/cache@v4
         with:
           path: |
             ~/.cargo/bin
