+++ Makefile.toml
@@ -466,7 +466,7 @@ args = [
     "-scheme",
     "LiveViewNativeCore",
     "-destination",
-    "name=Apple TV 4K (3rd generation)",
+    "name=Apple TV 4K (3rd generation),arch=x86_64",
     "-workspace",
     "${CARGO_MAKE_WORKSPACE_WORKING_DIRECTORY}/crates/core/liveview-native-core-swift/",
     "test"
