+++ crates/core/Cargo.toml
@@ -26,7 +26,7 @@ name = "liveview_native_core"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cranelift-entity = { version = "0.101" }
+cranelift-entity = { version = "0.104" }
 fixedbitset = { version = "0.4" }
 fxhash = { version = "0.2" }
 html5gum = { git = "https://github.com/liveviewnative/html5gum", branch = "lvn" }
@@ -34,7 +34,6 @@ intrusive-collections = { version = "0.9", features = ["nightly"] }
 indexmap = { version = "2.1", features = ["std"] }
 petgraph = { version = "0.6", default-features = false, features = ["graphmap"] }
 phoenix_channels_client = { git = "https://github.com/liveview-native/phoenix-channels-client", branch = "simlay/more-fixes-for-file-upload" }
-#phoenix_channels_client = { path = "../../../phoenix-channels-client/" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0" }
 smallstr = { version = "0.3", features = ["union"] }
