--- crates/core/tests/support/app/.gitignore
@@ -1,1 +1,0 @@
-integration_test
