--- crates/core/tests/support/app/Info.plist
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-  <dict>
-    <key>CFBundleExecutable</key>
-    <string>test</string>
-    <key>CFBundleVersion</key>
-    <string>0.0.1</string>
-    <key>CFBundleShortVersionString</key>
-    <string>0.0.1</string>
-    <key>CFBundleIdentifier</key>
-    <string>com.rust.tests</string>
-    <key>UIRequiredDeviceCapabilities</key>
-    <array>
-      <string>x86_64</string>
-    </array>
-  </dict>
-</plist>
