--- crates/core/tests/support/linker_shim.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-cc -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator.sdk "$@"
